Changed damage mechanic - damage is now array, that can be set as rescource - updated fireball to rescource - projectile spells are now rescource based - enemy damage is now rescource (will be further modified)

+++ entities/enemies/FireImp.cs
@@ -2,8 +2,8 @@ using System.Collections.Generic;
 
 public partial class FireImp : EnemyEntity
 {
-    private int _hp = 250;
-    public override int Hp { get => _hp; set => _hp = value; }
+    private double _hp = 250;
+    public override double Hp { get => _hp; set => _hp = value; }
     public override float Speed { get => 1.5f; }
 
     public override ElementalValues Resistances => new(new Dictionary<ElementType, int>
