+++ entities/static/Dummy.cs
@@ -3,7 +3,7 @@ using System;
 
 public partial class Dummy : BaseEntity
 {
-	private int _hp = 2000;
-	public override int Hp { get => _hp; set => _hp = value; }
+	private double _hp = 2000;
+	public override double Hp { get => _hp; set => _hp = value; }
 
 }
